Merge in KI_ATLAS/atlas from feature/ATLAS-2812-fix-hibernate-flush to master

* commit '655ff20cdb30fe226a8ed9f796c8d3ced166b819':
  ATLAS-2812: fix hibernate flush

+++ line-directory/src/main/java/ch/sbb/line/directory/service/SublineService.java
@@ -71,7 +71,9 @@ public class SublineService {
 
   @Transactional
   public void deleteById(Long id) {
-    SublineVersion sublineVersion = sublineVersionRepository.findById(id).orElseThrow(() -> new IdNotFoundException(id));
+    if (!sublineVersionRepository.existsById(id)) {
+      throw new IdNotFoundException(id);
+    }
     sublineVersionRepository.deleteById(id);
   }
